requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Local fallbacks (no cloud dependencies for basic functionality)
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
reportlab==4.0.7
langdetect==1.0.9

# Optional cloud services (will fallback if not available)
google-cloud-vision==3.4.4
google-cloud-translate==3.11.1
google-cloud-aiplatform==1.38.1
openai==1.3.7
